Update.004?
So I happened to find a fantastic tutorial, modified the code, deconstructed it to understand it, and then trimmed out all the excess junk to end up with this commit, which is an original template for making tabbed applications in as little as code as possible. And honestly, I think I did a pretty good job.